+++ readme.md
@@ -1,0 +1,7 @@
+# inMobile PHP API
+
+This is a fork of the official PHP api for inMobile, just made properly to be required via composer.
+
+## Requirements
+- php:  *>7.1*
+- composer installed
