+++ src/SimpleXMLExtended.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace MarkVesterskov\InMobile;
+
+class SimpleXMLExtended extends \SimpleXMLExtended
+{
+    public function addCData($cdata_text)
+    {
+        $node = dom_import_simplexml($this);
+        $no = $node->ownerDocument;
+        $node->appendChild($no->createCDATASection($cdata_text));
+    }
+}
